requirements.txt
pydantic>=2.5.0,<3.0.0
crewai==0.11.0
langchain>=0.1.13
langchain-community>=0.0.13
langchain-cohere>=0.0.1
cohere>=4.37
python-dotenv>=1.0.1
weaviate-client>=4.5.4

